Many changes before release

+++ app/src/main/java/com/vodovoz/app/data/vodovoz_service/model/cart/BottomCartDTO.kt
@@ -7,7 +7,7 @@ import com.squareup.moshi.Json
 @Keep
 data class BottomCartDTO(
     @Json(name = "ALLSUMA")
-    val ALLSUMA: Int?,
+    val ALLSUMA: Float?,
     @Json(name = "QUANTITY")
     val QUANTITY: Int?,
     @Json(name = "TOVAROV")
